+++ src/euroeval/dataset_configs/german.py
@@ -161,3 +161,12 @@ INCLUDE_DE_CONFIG = DatasetConfig(
     languages=[GERMAN],
     unofficial=True,
 )
+
+MULTILOKO_DE_CONFIG = DatasetConfig(
+    name="multiloko-de",
+    pretty_name="MultiLoKo-de",
+    source="EuroEval/multiloko-de-mini",
+    task=KNOW,
+    languages=[GERMAN],
+    unofficial=True,
+)