+++ src/euroeval/dataset_configs/italian.py
@@ -152,3 +152,12 @@ INCLUDE_IT_CONFIG = DatasetConfig(
     languages=[ITALIAN],
     unofficial=True,
 )
+
+MULTILOKO_IT_CONFIG = DatasetConfig(
+    name="multiloko-it",
+    pretty_name="MultiLoKo-it",
+    source="EuroEval/multiloko-it-mini",
+    task=KNOW,
+    languages=[ITALIAN],
+    unofficial=True,
+)