+++ src/euroeval/dataset_configs/spanish.py
@@ -161,3 +161,12 @@ INCLUDE_ES_CONFIG = DatasetConfig(
     languages=[SPANISH],
     unofficial=True,
 )
+
+MULTILOKO_ES_CONFIG = DatasetConfig(
+    name="multiloko-es",
+    pretty_name="MultiLoKo-es",
+    source="EuroEval/multiloko-es-mini",
+    task=KNOW,
+    languages=[SPANISH],
+    unofficial=True,
+)