+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   This was added by @harderj ✨
 - Added the new Danish linguistic acceptability dataset DaLA. It's marked as
   unofficial for now. This was added by @N-essuno ✨
+- Added the MultiLoKo multilingual local knowledge benchmark datasets for Dutch,
+  English, French, German, Italian, Portuguese, Spanish and Swedish. All datasets
+  are marked as `unofficial` for now.
 - Added the Schibsted front-page title and SEO title datasets, sourced from
   single newsrooms: `vg-front-title` features front-page titles from VG (Norwegian),
   and `svd-seo-title` features SEO titles from Svenska Dagbladet (Swedish). Both